+++ frontend/assets/styles/components/favorites.scss
@@ -3,8 +3,12 @@
   line-height: 14px;
 
   .fave-content {
-    padding: 12px 12px 12px 12px !important;
+    padding: 12px 5px 0 5px !important;
   }
+}
+
+.fave-date-container {
+  padding-bottom: 5px;
 }
 
 .fave-card:nth-child(odd) {
@@ -32,4 +36,8 @@
   display: flex;
   flex-direction: row;
   line-height: 24px;
+}
+
+.modal-button {
+  text-transform: capitalize;
 }
